remove extraneous html-macro stuff
